[frontend] implements quiz dialog component
