+++ frontend/public/locales/en/quiz.json
@@ -1,0 +1,98 @@
+{
+  "confirmation": {
+    "sure": "Are you sure you want to close this quiz? You will lose your progress.",
+    "no": "No, Return to Quiz",
+    "yes": "Yes, Close"
+  },
+  "navigation": {
+    "start": "Start The Quiz",
+    "previous": "Previous Question",
+    "next": "Next Question",
+    "submit": "Submit",
+    "close": "Close",
+    "title": "Retirement Ready Quiz",
+    "of": "of"
+  },
+  "landing": {
+    "p1": "Not sure you've checked all the boxes to retire? Take this quiz as a first step towards your retirement checklist.",
+    "p2": "Your answers will help identify government pensions and benefits you could receive. They will also guide you on what to consider to better prepare for retirement.",
+    "whatyouneed": "What you'll need",
+    "p3": "You don't need any special information to take this quiz. To create the most appropriate retirement checklist, you will need to provide information on the following:",
+    "legal": "<strong>legal status</strong> (such as Canadian citizen, Indian Status, or permanent resident)",
+    "residence": "<strong>residence history</strong> (number of years lived in Canada)",
+    "marital": "<strong>marital status</strong>",
+    "retirement": "<strong>retirement plan</strong> (when you plan on retiring, where you plan on living in retirement, etc.)",
+    "timetocomplete": "Time to complete survey",
+    "minutes": "3 minutes"
+  },
+  "questions": {
+    "question-1": {
+      "title": "When do you plan on retiring?",
+      "option-1": "Before Age 60",
+      "option-2": "Between ages 60 and 65",
+      "option-3": "At Age 65",
+      "option-4": "Between ages 65 and 70",
+      "option-5": "After Age 70",
+      "option-6": "Not sure"
+    },
+    "question-2": {
+      "question-1": {
+        "title": "Select all that apply to you. You are...",
+        "option-1": "Single",
+        "option-2": "Married or common-law",
+        "option-3": "Divorced or separated",
+        "option-4": "Widowed"
+      },
+      "question-2": {
+        "title": "Do you have children?",
+        "option-1": "Yes",
+        "option-2": "No"
+      }
+    },
+    "question-3": {
+      "title": "How financially prepared do you feel for retirement?",
+      "subtitle": "Do you feel confident that you have enough money to support your lifestyle and future plans in retirement?",
+      "option-1": "Very unprepared",
+      "option-2": "Unprepared",
+      "option-3": "Not sure",
+      "option-4": "Prepared",
+      "option-5": "Very Prepared"
+    },
+    "question-4": {
+      "title": "Where do you plan on living when you retire?",
+      "option-1": "In Canada",
+      "option-2": "In Canada MORE Than 6 Months a Year",
+      "option-3": "In Canada LESS than 6 months a year",
+      "option-4": "Outside Canada",
+      "option-5": "Not Sure",
+      "option-2-label": "In Canada more than 6 months a year",
+      "option-3-label": "In Canada more than 6 months a year"
+    },
+    "question-5": {
+      "title": "Do you plan to earn money while receiving retirement benefits",
+      "subtitle": "Some people choose to work or earn money in retirement. This can affect your monthly pension amounts.",
+      "option-1": "No",
+      "option-2": "Yes",
+      "option-3": "Not sure"
+    },
+    "question-6": {
+      "title": "When you retire, what will your legal status be in Canada?",
+      "option-1": "Canadian Citizen or Permanent Resident",
+      "option-2": "Métis, Inuit, or Member of First Nation",
+      "option-3": "Sponsored Resident (less than 10 years in Canada)",
+      "option-4": "Visitor, Student, or Temporary Worker"
+    },
+    "question-7": {
+      "title": "From age 18, how many years will you have lived in Canada when you apply for retirement benefits?",
+      "option-1": "40 years or more",
+      "option-2": "10 to 39 years",
+      "option-3": "Less than 10 years",
+      "option-4": "Not sure"
+    },
+    "question-8": {
+      "title": "Are you receiving CPP-Disability benefits?",
+      "option-1": "Yes",
+      "option-2": "No"
+    }
+  }
+}
