+++ frontend/src/components/quiz/questions/Question1.tsx
@@ -8,18 +8,18 @@ type QuestionProps = {
   setFieldValue: (field: string, value: any, shouldValidate?: boolean) => void
 }
 
-const Question4 = ({ values, setFieldValue }: QuestionProps) => {
-  let { t } = useTranslation('learn')
+export const Question1 = ({ values, setFieldValue }: QuestionProps) => {
+  const { t } = useTranslation('quiz')
   const [value, setValue] = React.useState('')
 
   const handleChange = (event: React.MouseEvent<HTMLElement>, answerId: string) => {
     setValue(answerId)
-    setFieldValue('retirementTimeframe', values['retirementTimeframe'] === answerId ? '' : answerId ?? '')
+    setFieldValue('retirementAge', values['retirementAge'] === answerId ? '' : answerId ?? '')
   }
 
   return (
     <div>
-      <h5 className="font-display text-2xl font-light">{t('quiz.questions.question-4.title')}</h5>
+      <h5 className="font-display text-2xl font-light">{t('questions.question-1.title')}</h5>
       <ToggleButtonGroup
         orientation="vertical"
         exclusive
@@ -42,48 +42,54 @@ const Question4 = ({ values, setFieldValue }: QuestionProps) => {
         }}
       >
         <ToggleButton
-          value="canada-ft"
-          aria-label={t('quiz.questions.question-4.option-1')}
+          value="pre-60"
+          aria-label={t('questions.question-1.option-1')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['retirementTimeframe'] === 'canada-ft'}
+          selected={values['retirementAge'] === 'pre-60'}
         >
-          {t('quiz.questions.question-4.option-1')}
+          {t('questions.question-1.option-1')}
         </ToggleButton>
         <ToggleButton
-          value="canada-pt-60-or-more"
-          aria-label={t('quiz.questions.question-4.option-2')}
+          value="between-60-and-65"
+          aria-label={t('questions.question-1.option-2')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['retirementTimeframe'] === 'canada-pt-60-or-more'}
+          selected={values['retirementAge'] === 'between-60-and-65'}
         >
-          {t('quiz.questions.question-4.option-2')}
+          {t('questions.question-1.option-2')}
         </ToggleButton>
         <ToggleButton
-          value="canada-pt-less-than-60"
-          aria-label={t('quiz.questions.question-4.option-3')}
+          value="at-65"
+          aria-label={t('questions.question-1.option-3')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['retirementTimeframe'] === 'canada-pt-less-than-60'}
+          selected={values['retirementAge'] === 'at-65'}
         >
-          {t('quiz.questions.question-4.option-3')}
+          {t('questions.question-1.option-3')}
         </ToggleButton>
         <ToggleButton
-          value="outside-canada"
-          aria-label={t('quiz.questions.question-4.option-4')}
+          value="between-65-and-70"
+          aria-label={t('questions.question-1.option-4')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['retirementTimeframe'] === 'outside-canada'}
+          selected={values['retirementAge'] === 'between-65-and-70'}
         >
-          {t('quiz.questions.question-4.option-4')}
+          {t('questions.question-1.option-4')}
         </ToggleButton>
         <ToggleButton
-          value="unsure-retirement-living"
-          aria-label={t('quiz.questions.question-4.option-5')}
+          value="after-70"
+          aria-label={t('questions.question-1.option-5')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['retirementTimeframe'] === 'unsure-retirement-living'}
+          selected={values['retirementAge'] === 'after-70'}
         >
-          {t('quiz.questions.question-4.option-5')}
+          {t('questions.question-1.option-5')}
+        </ToggleButton>
+        <ToggleButton
+          value="unsure-retirement-age"
+          aria-label={t('questions.question-1.option-6')}
+          className="my-4 font-display text-base font-bold normal-case"
+          selected={values['retirementAge'] === 'unsure-retirement-age'}
+        >
+          {t('questions.question-1.option-6')}
         </ToggleButton>
       </ToggleButtonGroup>
     </div>
   )
 }
-
-export default Question4
