+++ frontend/src/components/quiz/questions/Question3.tsx
@@ -8,18 +8,19 @@ type QuestionProps = {
   setFieldValue: (field: string, value: any, shouldValidate?: boolean) => void
 }
 
-const Question6 = ({ values, setFieldValue }: QuestionProps) => {
-  let { t } = useTranslation('learn')
+export const Question3 = ({ values, setFieldValue }: QuestionProps) => {
+  const { t } = useTranslation('quiz')
   const [value, setValue] = React.useState('')
 
   const handleChange = (event: React.MouseEvent<HTMLElement>, answerId: string) => {
     setValue(answerId)
-    setFieldValue('legalStatus', values['legalStatus'] === answerId ? '' : answerId ?? '')
+    setFieldValue('financialPreparedness', values['financialPreparedness'] === answerId ? '' : answerId ?? '')
   }
 
   return (
     <div>
-      <h5 className="font-display text-2xl font-light">{t('quiz.questions.question-6.title')}</h5>
+      <h5 className="font-display text-2xl font-light">{t('questions.question-3.title')}</h5>
+      <p className="mt-5 font-display text-sm font-light">{t('questions.question-3.subtitle')}</p>
       <ToggleButtonGroup
         orientation="vertical"
         exclusive
@@ -42,40 +43,46 @@ const Question6 = ({ values, setFieldValue }: QuestionProps) => {
         }}
       >
         <ToggleButton
-          value="status-citizen"
-          aria-label={t('quiz.questions.question-6.option-1')}
+          value="very-unprepared"
+          aria-label={t('questions.question-3.option-1')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['legalStatus'] === 'status-citizen'}
+          selected={values['financialPreparedness'] === 'very-unprepared'}
         >
-          {t('quiz.questions.question-6.option-1')}
+          {t('questions.question-3.option-1')}
         </ToggleButton>
         <ToggleButton
-          value="status-first-nation"
-          aria-label={t('quiz.questions.question-6.option-2')}
+          value="unprepared"
+          aria-label={t('questions.question-3.option-2')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['legalStatus'] === 'status-first-nation'}
+          selected={values['financialPreparedness'] === 'unprepared'}
         >
-          {t('quiz.questions.question-6.option-2')}
+          {t('questions.question-3.option-2')}
         </ToggleButton>
         <ToggleButton
-          value="status-sponsored"
-          aria-label={t('quiz.questions.question-6.option-3')}
+          value="unsure-preparedness"
+          aria-label={t('questions.question-3.option-3')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['legalStatus'] === 'status-sponsored'}
+          selected={values['financialPreparedness'] === 'unsure-preparedness'}
         >
-          {t('quiz.questions.question-6.option-3')}
+          {t('questions.question-3.option-3')}
         </ToggleButton>
         <ToggleButton
-          value="status-other"
-          aria-label={t('quiz.questions.question-6.option-4')}
+          value="prepared"
+          aria-label={t('questions.question-3.option-4')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['legalStatus'] === 'status-other'}
+          selected={values['financialPreparedness'] === 'prepared'}
         >
-          {t('quiz.questions.question-6.option-4')}
+          {t('questions.question-3.option-4')}
+        </ToggleButton>
+        <ToggleButton
+          value="very-prepared"
+          aria-label={t('questions.question-3.option-5')}
+          className="my-4 font-display text-base font-bold normal-case"
+          selected={values['financialPreparedness'] === 'very-prepared'}
+        >
+          {t('questions.question-3.option-5')}
         </ToggleButton>
       </ToggleButtonGroup>
     </div>
   )
 }
-
-export default Question6
