+++ frontend/src/components/quiz/questions/Question4.tsx
@@ -8,19 +8,18 @@ type QuestionProps = {
   setFieldValue: (field: string, value: any, shouldValidate?: boolean) => void
 }
 
-const Question3 = ({ values, setFieldValue }: QuestionProps) => {
-  let { t } = useTranslation('learn')
+export const Question4 = ({ values, setFieldValue }: QuestionProps) => {
+  const { t } = useTranslation('quiz')
   const [value, setValue] = React.useState('')
 
   const handleChange = (event: React.MouseEvent<HTMLElement>, answerId: string) => {
     setValue(answerId)
-    setFieldValue('financialPreparedness', values['financialPreparedness'] === answerId ? '' : answerId ?? '')
+    setFieldValue('retirementTimeframe', values['retirementTimeframe'] === answerId ? '' : answerId ?? '')
   }
 
   return (
     <div>
-      <h5 className="font-display text-2xl font-light">{t('quiz.questions.question-3.title')}</h5>
-      <p className="mt-5 font-display text-sm font-light">{t('quiz.questions.question-3.subtitle')}</p>
+      <h5 className="font-display text-2xl font-light">{t('questions.question-4.title')}</h5>
       <ToggleButtonGroup
         orientation="vertical"
         exclusive
@@ -43,48 +42,46 @@ const Question3 = ({ values, setFieldValue }: QuestionProps) => {
         }}
       >
         <ToggleButton
-          value="very-unprepared"
-          aria-label={t('quiz.questions.question-3.option-1')}
+          value="canada-ft"
+          aria-label={t('questions.question-4.option-1')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['financialPreparedness'] === 'very-unprepared'}
+          selected={values['retirementTimeframe'] === 'canada-ft'}
         >
-          {t('quiz.questions.question-3.option-1')}
+          {t('questions.question-4.option-1')}
         </ToggleButton>
         <ToggleButton
-          value="unprepared"
-          aria-label={t('quiz.questions.question-3.option-2')}
+          value="canada-pt-60-or-more"
+          aria-label={t('questions.question-4.option-2')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['financialPreparedness'] === 'unprepared'}
+          selected={values['retirementTimeframe'] === 'canada-pt-60-or-more'}
         >
-          {t('quiz.questions.question-3.option-2')}
+          {t('questions.question-4.option-2')}
         </ToggleButton>
         <ToggleButton
-          value="unsure-preparedness"
-          aria-label={t('quiz.questions.question-3.option-3')}
+          value="canada-pt-less-than-60"
+          aria-label={t('questions.question-4.option-3')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['financialPreparedness'] === 'unsure-preparedness'}
+          selected={values['retirementTimeframe'] === 'canada-pt-less-than-60'}
         >
-          {t('quiz.questions.question-3.option-3')}
+          {t('questions.question-4.option-3')}
         </ToggleButton>
         <ToggleButton
-          value="prepared"
-          aria-label={t('quiz.questions.question-3.option-4')}
+          value="outside-canada"
+          aria-label={t('questions.question-4.option-4')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['financialPreparedness'] === 'prepared'}
+          selected={values['retirementTimeframe'] === 'outside-canada'}
         >
-          {t('quiz.questions.question-3.option-4')}
+          {t('questions.question-4.option-4')}
         </ToggleButton>
         <ToggleButton
-          value="very-prepared"
-          aria-label={t('quiz.questions.question-3.option-5')}
+          value="unsure-retirement-living"
+          aria-label={t('questions.question-4.option-5')}
           className="my-4 font-display text-base font-bold normal-case"
-          selected={values['financialPreparedness'] === 'very-prepared'}
+          selected={values['retirementTimeframe'] === 'unsure-retirement-living'}
         >
-          {t('quiz.questions.question-3.option-5')}
+          {t('questions.question-4.option-5')}
         </ToggleButton>
       </ToggleButtonGroup>
     </div>
   )
 }
-
-export default Question3
