+++ frontend/src/pages/learn/index.tsx
@@ -1,186 +1,31 @@
-import { FC, useEffect, useState } from 'react'
+import { FC, useState } from 'react'
 
-import CloseIcon from '@mui/icons-material/Close'
-import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline'
-import {
-  Button,
-  Card,
-  CardActionArea,
-  CardContent,
-  CardMedia,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  LinearProgress,
-  useMediaQuery,
-  useTheme,
-} from '@mui/material'
-import { useFormikWizard } from 'formik-wizard-form'
-import { compact } from 'lodash'
+import { Button, Card, CardActionArea, CardContent, CardMedia } from '@mui/material'
 import { GetServerSideProps } from 'next'
 import { useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { NextSeo } from 'next-seo'
 import Image from 'next/image'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 
 import Layout from '../../components/Layout'
-import QuizLandingPage from '../../components/QuizLandingPage'
-import Question1 from '../../components/questions/Question1'
-import Question2 from '../../components/questions/Question2'
-import Question3 from '../../components/questions/Question3'
-import Question4 from '../../components/questions/Question4'
-import Question5 from '../../components/questions/Question5'
-import Question6 from '../../components/questions/Question6'
-import Question7 from '../../components/questions/Question7'
-import Question8 from '../../components/questions/Question8'
-import { useQuizData } from '../../lib/hooks/useQuizData'
-import { useRemoveQuizData } from '../../lib/hooks/useRemoveQuizData'
-import { useSetQuizData } from '../../lib/hooks/useSetQuizData'
-import { Filters } from '../quiz/tasks/[[...filters]]'
-
-// TODO: map form values to "answer-id" from locales/(en/fr)/quiz/tasks/task-list.json once it has been finalized
-export interface QuizFormState extends Record<string, string> {
-  retirementAge: string
-  single: string
-  marriedOrCommonLaw: string
-  divorcedOrSeparated: string
-  widowed: string
-  hasChildren: string
-  financialPreparedness: string
-  retirementTimeframe: string
-  hasExtraIncome: string
-  legalStatus: string
-  yearsInCanada: string
-  hasCppDisabilityBenefits: string
-}
-
-const defaultFormValues: QuizFormState = {
-  retirementAge: '',
-  single: '',
-  marriedOrCommonLaw: '',
-  divorcedOrSeparated: '',
-  widowed: '',
-  hasChildren: '',
-  financialPreparedness: '',
-  retirementTimeframe: '',
-  hasExtraIncome: '',
-  legalStatus: '',
-  yearsInCanada: '',
-  hasCppDisabilityBenefits: '',
-}
-
-export interface QuizConfirmationProps {
-  sureText: string
-  noText: string
-  yesText: string
-  handleConfirmationCancel: () => void
-  handleCloseModal: () => void
-}
-
-const QuizConfirmation: FC<QuizConfirmationProps> = ({
-  handleConfirmationCancel,
-  handleCloseModal,
-  sureText,
-  noText,
-  yesText,
-}) => {
-  return (
-    <>
-      <DialogContent>
-        <div className="mb-10 text-center">
-          <ErrorOutlineIcon className="text-9xl text-red-dark" />
-        </div>
-        <p>{sureText}</p>
-      </DialogContent>
-      <DialogActions className="block">
-        <div className="grid gap-2 md:grid-cols-2 md:gap-6">
-          <Button onClick={handleConfirmationCancel} variant="outlined" size="large" fullWidth>
-            {noText}
-          </Button>
-          <Button onClick={handleCloseModal} size="large" fullWidth>
-            {yesText}
-          </Button>
-        </div>
-      </DialogActions>
-    </>
-  )
-}
+import { QuizDialog } from '../../components/quiz/QuizDialog'
 
 const Learn: FC = () => {
-  const router = useRouter()
   const { t } = useTranslation('learn')
-  const theme = useTheme()
-  const fullScreen = useMediaQuery(theme.breakpoints.down('md'))
-
-  const formikWizard = useFormikWizard({
-    initialValues: defaultFormValues,
-    onSubmit: (values) => {
-      // Encodes a js object as a url-safe base64 string.
-      const filters: Filters = { answers: compact(Object.values<string>(values)) }
-      const encodedFilters = encodeURIComponent(window.btoa(JSON.stringify(filters)))
-      router.push(`/quiz/tasks/${encodedFilters}`)
-    },
-    validateOnNext: true,
-    activeStepIndex: 0,
-    steps: [
-      { component: QuizLandingPage },
-      { component: Question1 },
-      { component: Question2 },
-      { component: Question3 },
-      { component: Question4 },
-      { component: Question5 },
-      { component: Question6 },
-      { component: Question7 },
-      { component: Question8 },
-    ],
-  })
-
-  useQuizData({
-    onSuccess: (data) => {
-      if (data) {
-        // override initial form values with stored values
-        formikWizard.setValues({ ...defaultFormValues, ...data })
-      }
-    },
-  })
-
-  const { mutate: removeQuizData } = useRemoveQuizData()
-  const { mutate: setQuizData } = useSetQuizData()
+  const [quizDialogOpen, setQuizDialogOpen] = useState(false)
 
   const sections = t<string, { cards: any[] }[]>('sections', {
     returnObjects: true,
   })
 
-  const [isModalOpen, setIsModalOpen] = useState(false)
-  const [showConfirmation, setShowConfirmation] = useState(false)
-
-  const handleOpenModal = () => {
-    setIsModalOpen(true)
+  const handleOnQuizDialogTriggerClick = () => {
+    setQuizDialogOpen(true)
   }
 
-  const handleCloseModal = () => {
-    if (showConfirmation) {
-      removeQuizData()
-      setIsModalOpen(false)
-      // 1 second timeout otherwise dialog shows quiz content and close
-      setTimeout(function () {
-        setShowConfirmation(false)
-      }, 1000)
-    } else {
-      setShowConfirmation(true)
-    }
+  const handleOnQuizDialogClose = () => {
+    setQuizDialogOpen(false)
   }
-
-  const handleConfirmationCancel = () => {
-    setShowConfirmation(false)
-  }
-
-  useEffect(() => {
-    setQuizData(formikWizard.values as QuizFormState)
-  }, [formikWizard.values, setQuizData])
 
   return (
     <Layout
@@ -201,97 +46,12 @@ const Learn: FC = () => {
           <div className="my-2 px-6 pb-10 text-center md:w-2/3 md:pb-10 md:pl-12 md:pt-10 md:text-left lg:w-4/5 lg:pb-4">
             <h2 className="mb-4 text-left font-display text-5xl font-bold text-primary-700">{t('banner.title')}</h2>
             <p className="pb-4 text-left text-lg font-normal md:w-4/5">{t('banner.text')}</p>
-            <Button size="large" onClick={handleOpenModal}>
+            <Button id="quiz-dialog-trigger" size="large" onClick={handleOnQuizDialogTriggerClick}>
               {t('banner.quiz')}
             </Button>
           </div>
         </div>
       </section>
-
-      <Dialog
-        onClose={handleCloseModal}
-        open={isModalOpen}
-        aria-describedby="QuizModal-header"
-        scroll="body"
-        fullScreen={fullScreen}
-        maxWidth="md"
-        fullWidth
-      >
-        {showConfirmation ? (
-          <QuizConfirmation
-            handleCloseModal={handleCloseModal}
-            handleConfirmationCancel={handleConfirmationCancel}
-            sureText={t('quiz.confirmation.sure')}
-            yesText={t('quiz.confirmation.yes')}
-            noText={t('quiz.confirmation.no')}
-          />
-        ) : (
-          <>
-            <div className="flex min-h-[850px] flex-col">
-              <DialogTitle className="text-right">
-                <Button variant="text" onClick={handleCloseModal} startIcon={<CloseIcon />} size="large">
-                  {t('quiz.navigation.close')}
-                </Button>
-              </DialogTitle>
-              <DialogContent className="flex flex-col">
-                <h2 className="mb-8 font-display text-2xl font-medium md:mb-16 md:rounded-3xl md:bg-[#f5f5f5] md:p-6 md:text-4xl md:text-primary-700">
-                  {t('quiz.navigation.title')}
-                </h2>
-                <div className="mb-5">{formikWizard.renderComponent()}</div>
-                {!formikWizard.isFirstStep && (
-                  <div className="mt-auto">
-                    <LinearProgress
-                      variant="determinate"
-                      value={((formikWizard.currentStepIndex ?? 0) / 8) * 100}
-                      aria-labelledby="progress-label"
-                      className="my-2"
-                    />
-                    <p id="progress-label" className="m-0 text-center">
-                      {formikWizard.currentStepIndex} {t('quiz.navigation.of')} 8
-                    </p>
-                  </div>
-                )}
-              </DialogContent>
-              <DialogActions className="block">
-                {formikWizard.isFirstStep ? (
-                  <Button
-                    onClick={formikWizard.handleNext}
-                    disabled={formikWizard.isNextDisabled}
-                    size="large"
-                    fullWidth
-                    autoFocus
-                  >
-                    {t('quiz.navigation.start')}
-                  </Button>
-                ) : (
-                  <>
-                    <div className="grid gap-2 md:grid-cols-2 md:gap-6">
-                      <Button
-                        onClick={formikWizard.handlePrev}
-                        disabled={formikWizard.isPrevDisabled}
-                        size="large"
-                        fullWidth
-                        variant="outlined"
-                      >
-                        {t('quiz.navigation.previous')}
-                      </Button>
-                      <Button
-                        onClick={formikWizard.handleNext}
-                        disabled={formikWizard.isNextDisabled}
-                        size="large"
-                        fullWidth
-                      >
-                        {formikWizard.isLastStep ? t('quiz.navigation.submit') : t('quiz.navigation.next')}
-                      </Button>
-                    </div>
-                  </>
-                )}
-              </DialogActions>
-            </div>
-          </>
-        )}
-      </Dialog>
-
       <section>
         {sections.map((section, index) => (
           <div key={index}>
@@ -328,6 +88,7 @@ const Learn: FC = () => {
           </div>
         ))}
       </section>
+      <QuizDialog open={quizDialogOpen} onClose={handleOnQuizDialogClose} />
     </Layout>
   )
 }
@@ -335,7 +96,7 @@ const Learn: FC = () => {
 export const getServerSideProps: GetServerSideProps = async ({ locale }) => {
   return {
     props: {
-      ...(await serverSideTranslations(locale ?? 'default', ['common', 'learn'])),
+      ...(await serverSideTranslations(locale ?? 'default', ['common', 'learn', 'quiz'])),
     },
   }
 }
